added function filter_name_rows
This function has an argument which is a single row from a generation dataset. This function then matches the name from that row with the same name in the global stats-dataset, so that we end up returning a new single row which has all the stats which the generation-dataset does not have.