added function get_pokemon_stats
We use the argument generation_dropdown to read the corresponding generation dataset. We then use the argument strongest_weakest_dropdown to fnd either the strongest or weakest pokemon in the selected generation dataset. We return the strongest/weakest pokemon row data as a list of dictionaries (with 1 element) which we will then use for our dash.table.DataTable data-property