added functions pokemon_stats_graph and helper_pokemon_stats_graph
helper_ is a helper function which is used to avoid duplicationg of code in pokemon_stats_graph. Together these function will return a new row of number stats of the resulting pokemon. Is it also a dictionary and will be used in the comparison-graph of average stats vs resulting pokemon stats